requirements.txt
python-magic
termcolor
